+++ CHANGELOG.md
@@ -2016,6 +2016,9 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We fixed an issue where opening the menu 'Library properties' marked the library as modified [#6451](https://github.com/JabRef/jabref/issues/6451)
 - We fixed an issue when importing resulted in an exception [#7343](https://github.com/JabRef/jabref/issues/7343)
 - We fixed an issue where the field in the Field formatter dropdown selection were sorted in random order. [#7710](https://github.com/JabRef/jabref/issues/7710)
+- We fixed an issue where opening a group chat window in one library could show the chat history from a group in a different library, given the two groups had the same name[#14641]
+
+
 
 ### Removed